+++ project/src/pages/bookmark/Bookmark.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import tw from "tailwind-styled-components";
 import Filter from "../../components/filter/Filter";
-import { useRef, useState, useEffect } from "react";
+import useUnlimitScroll from "../../hooks/useUnlimitScroll";
 import { useSelector, useDispatch } from "react-redux";
 import { RootState } from "../../modules";
 import { closeModal } from "../../modules/hamburgerModalSlice";
@@ -9,25 +9,12 @@ import Categories from "../../types/categories";
 import ProductCard from "../../components/productCard/ProductCard";
 
 const Bookmark : FC  = () => {
-  const scrollRef = useRef<number | null>(null);
   const dispatch = useDispatch();
   const productList = useSelector((state: RootState)  => state.productList.products);
   const category = useSelector((state: RootState) => state.filterList.category);
 
-  const defaultPage: number = 8;
-  const [page, setPage] = useState(defaultPage);
-
-  useEffect(() => {
-    const scrollListener = () => {
-      scrollRef.current = window.scrollY;
-      const scrollPosition = window.innerHeight + document.documentElement.scrollTop;
-      const scrollHeight = document.documentElement.scrollHeight;
-      if (scrollPosition === scrollHeight) setPage(page + defaultPage);
-    };
-
-    window.addEventListener("scroll", scrollListener);
-    return () => window.removeEventListener("scroll", scrollListener);
-  }, [page, setPage]);
+  const DEFAULTPAGE: number = 8;
+  const page = useUnlimitScroll(DEFAULTPAGE, productList.length);
 
   switch(category){
     case Categories.total :
